pyproject.toml
[project]
name = "khmer-video-editor"
version = "0.1.0"
requires-python = ">=3.9, <=3.11"
dependencies = [
    "googletrans==4.0.0-rc1",
    "openai-whisper",         # speech‑to‑text model
    "setuptools-rust",        # build support for Rust extensions
    "torch",                  # PyTorch core (CUDA 12.6 build — see [tool.uv.sources])
    "torchvision",            # computer vision utilities
    "torchaudio",             # audio processing utilities
    "edge-tts",               # Khmer TTS (Microsoft Edge Neural voices)
    "pyside6",                # Qt6 UI framework
    "ffmpeg-python",          # FFmpeg Python bindings for export
]

# -------------------\--------------------------------------------------------
# uv: pin torch/torchvision/torchaudio to the CUDA 12.6 PyTorch index so
# that `uv sync` installs GPU-enabled wheels instead of the default CPU build.
# Change the index name/url below if you use a different CUDA version:
#   CUDA 11.8 → https://download.pytorch.org/whl/cu118
#   CUDA 12.1 → https://download.pytorch.org/whl/cu121
#   CUDA 12.4 → https://download.pytorch.org/whl/cu124
#   CUDA 12.6 → https://download.pytorch.org/whl/cu126  (current)
# ---------------------------------------------------------------------------
[tool.uv.sources]
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }
torchaudio = { index = "pytorch-cu126" }

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
